+++ src/Device.ts
@@ -1,4 +1,7 @@
+import { DeviceClient } from ".";
+
 export default interface Device {
   id: string;
   type: 'emulator' | 'device' | 'offline';
+  getClient: () => DeviceClient;
 }