+++ src/adb/command/host-transport/index.ts
@@ -30,3 +30,4 @@ export { default as PsCommand, PsEntry } from './ps';
 export { default as ServicesListCommand, AdbServiceInfo, KnownServices } from './servicesList';
 export { default as ServiceCheckCommand } from './serviceCheck';
 export { default as ServiceCallCommand } from './serviceCall';
+export { default as IpRouteCommand } from './ip_route';