+++ test/mock/client.ts
@@ -1,0 +1,7 @@
+import Client from '../../src/adb/client';
+
+export default class MockClient extends Client {
+    constructor() {
+        super();
+    }
+}